safe singleton management with a lifetime explicit control
When C++ code gets used as a underlying layer for scripting languages like JS, singleton management has became a critical part of the code stability. Here is the example how to create a singleton storage for a safe singleton management with a lifetime explicit control of the singleton instances for different types.